Fill in details for elevator based on latest designs
This exercise exposed some new (to me) techniques for controlling the
update rate of data coming out of the Talons. This was employed in other
areas to make sure we are getting fresh values for the data that is
important and accepting older data for data which is not important to
our robot.  Slowing down the data rate of less important data helps
to offset the bandwidth cost of increasing data rate of the important
stuff.